requirements.txt
asgiref==3.8.1
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-admin-autocomplete-filter==0.7.1
django-cors-headers==4.4.0
django-money==3.5.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
environs==11.0.0
idna==3.8
marshmallow==3.22.0
packaging==24.1
phonenumberslite==8.13.44
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-moneyed==3.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
